ai-llm-course/requirements.txt
openai>=1.0.0
tiktoken>=0.5.0
python-dotenv>=1.0.0
langchain>=0.2.0
langchain-openai>=0.1.0
chromadb>=0.4.0
faiss-cpu>=1.7.0
rich>=13.0.0
